Backend/requirements.txt
beautifulsoup4
fastapi
langchain
langchain-community
langchain-core
numpy
ollama
pandas
pydantic
python-multipart
PyYAML
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
uvicorn
yfinance